refactor: split main.py into commands/ submodules
