+++ src/pyimgtag/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Subcommand handler modules for pyimgtag."""
